tests: fix constructor WeatherDataTest
OK, the next issue i found based on the test case, is, that the AllArgs test are still failed. Base on the first analyse it seems to be very clear. The assert statements contains hard coded zero as expected value. So i changed the method signature and pass all expected values to the assert method. Also i had to adjust the test cases again. but now i hope it will be run