reader: fix class loader access for resource file
In same cases the files could not be readed, because it's unknown resources inside the system class loader. So i changed it and use the class loader from it's class thats loaded the ClassLoaderReader. So i have to adjust each resource access. So i add a slash as prefix for each resource literal.